feat(altiore/form): better usage of altiore/form
