+++ src/index.js
@@ -5,7 +5,6 @@ import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  // React.StrictMode blocking usage altiore/form, because forms are not controlled
+  <App />
 );
